Move constructors, pipe, unique
