+++ python/cudf/cudf/tests/groupby/test_constructors.py
@@ -1,0 +1,16 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.
+import pandas as pd
+import pytest
+
+import cudf
+from cudf.testing._utils import assert_exceptions_equal
+
+
+@pytest.mark.parametrize("data", [{"a": [1, 2]}, {"a": [1, 2], "b": [2, 3]}])
+def test_groupby_nonempty_no_keys(data):
+    pdf = pd.DataFrame(data)
+    gdf = cudf.from_pandas(pdf)
+    assert_exceptions_equal(
+        lambda: pdf.groupby([]),
+        lambda: gdf.groupby([]),
+    )
